requirements.txt
yt-dlp
openai
together
pysrt
python-dotenv
